refactor: noslow
