+++ src/main/kotlin/net/ccbluex/liquidbounce/features/fakelag/FakeLag.kt
@@ -28,9 +28,9 @@ import net.ccbluex.liquidbounce.event.repeatable
 import net.ccbluex.liquidbounce.features.module.modules.combat.ModuleFakeLag
 import net.ccbluex.liquidbounce.features.module.modules.exploit.ModuleClickTp
 import net.ccbluex.liquidbounce.features.module.modules.movement.ModuleInventoryMove
-import net.ccbluex.liquidbounce.features.module.modules.movement.ModuleNoSlow
 import net.ccbluex.liquidbounce.features.module.modules.movement.autododge.ModuleAutoDodge
 import net.ccbluex.liquidbounce.features.module.modules.movement.fly.modes.specific.FlyNcpClip
+import net.ccbluex.liquidbounce.features.module.modules.movement.noslow.modes.blocking.NoSlowBlockingBlink
 import net.ccbluex.liquidbounce.features.module.modules.player.ModuleAntiVoid
 import net.ccbluex.liquidbounce.features.module.modules.player.ModuleBlink
 import net.ccbluex.liquidbounce.features.module.modules.player.nofall.modes.NoFallBlink
@@ -85,7 +85,7 @@ object FakeLag : Listenable {
             return LagResult.QUEUE
         }
 
-        ModuleNoSlow.Block.Blink.shouldLag(packet)?.let {
+        NoSlowBlockingBlink.shouldLag(packet)?.let {
             return it
         }
